Start of readding document creation and displaying

+++ module/UnpApplication/src/Controller/FileController.php
@@ -31,9 +31,7 @@ class FileController extends BaseController{
   /**
    * Sends the requested image file with the appropriate content-type, in order to get browsers to display it like
    * a really downloaded image in a publicly accessible folder. The benefit is, that we can check the permissions
-   * of the user here.
-   * 
-   * @todo secure the images
+   * of the user here, which wouldn't be possible in a public folder.
    */
   public function viewAction(){
     $id = $this->params('id');
@@ -43,18 +41,11 @@ class FileController extends BaseController{
       $file = $this->em->getRepository('\UnpCommon\Model\File')->findOneById($id);
 
       if($file && $file->isImage() && is_readable($file->getFullPath())){
-        /*
-          $permission = $this->_em->getRepository('\UnpCommon\Model\ModelPermission')->findOneBy(array('type'=>'file', 'action'=>'read', 'base'=>$file));
-          if(!Zend_Registry::getInstance()->user->getRole()->hasPermission($permission)){
-          $response->setStatusCode(403);
-          }
-         */
-
         $response->getHeaders()
-                ->addHeaderLine('Expires', '', true)
-                ->addHeaderLine('Cache-Control', 'private', true)
+                ->addHeaderLine('Expires', '')
+                ->addHeaderLine('Cache-Control', 'private')
                 ->addHeaderLine('Cache-Control', 'max-age=360000')
-                ->addHeaderLine('Pragma', '', true)
+                ->addHeaderLine('Pragma', '')
                 ->addHeaderLine('Content-type', 'image/' . $file->getExtension());
         $content = file_get_contents($file->getFullPath());
         $response->setContent($content);
@@ -77,11 +68,6 @@ class FileController extends BaseController{
     if(!empty($id)){
       $file = $this->em->getRepository('\UnpCommon\Model\File')->findOneById($id);
       if($file && is_readable($file->getFullPath())){
-        /* $permission = $this->_em->getRepository('Application_Model_ModelPermission')->findOneBy(array('type'=>'file', 'action'=>'read', 'base'=>$file));
-          if(!Zend_Registry::getInstance()->user->getRole()->hasPermission($permission)){
-          $this->redirectToLastPage(true);
-          }
-         */
 
         $response->getHeaders()
                 ->addHeaderLine('Cache-Control: no-cache, must-revalidate')
@@ -102,25 +88,22 @@ class FileController extends BaseController{
 
   /**
    * Deletes the file specified by the id parameter. 
+   * 
+   * @todo port to zf2
    */
   public function deleteAction(){
     $id = $this->params('id');
     $response = $this->getResponse();
-    $translator = $this->getServiceLocator()->get('translator');
-    //$this->flashMessenger()->setNamespace('success')->addMessage($translator->translate('The case was created successfully.'));
-
+    
     if(!empty($id)){
       $file = $this->em->getRepository('\UnpCommon\Model\File')->findOneById($id);
-      if($file){
-        /*
-          $permission = $this->em->getRepository('Application_Model_ModelPermission')->findOneBy(array('type'=>'file', 'action'=>'delete', 'base'=>$file));
-          if(!Zend_Registry::getInstance()->user->getRole()->hasPermission($permission)){
-          $this->redirectToLastPage(true);
-          } */
+      if($file && is_readable($file->getFullPath())){
+
         // remove file from file system
         $localPath = $file->getFullPath();
         $deleted = unlink($localPath);
-        if($deleted || !is_readable($localPath)){
+        if($deleted){
+        /*
           // set removed state in database record
           $file->remove();
 
@@ -148,19 +131,20 @@ class FileController extends BaseController{
           $guest = $registry->entitymanager->getRepository('Application_Model_User')->findOneById($guestId->getValue());
           $guest->removeFile($file);
           $this->em->persist($guest);
+         */
 
           $this->em->persist($file);
           $this->em->flush();
-          $this->flashMessenger()->setNamespace('success')->addMessage($translator->translate('The file was deleted successfully.'));
+          $this->flashMessenger()->setNamespace('success')->addMessage($this->getTranslator()->translate(sprintf('The file %s was deleted successfully.', $file->getFilename())));
         }else{
-          $this->flashMessenger()->setNamespace('error')->addMessage($translator->translate('We are sorry, but the file could not be deleted.'));
+          $this->flashMessenger()->setNamespace('error')->addMessage($this->getTranslator()->translate('We are sorry, but the file could not be deleted.'));
         }
       }else{
-        $this->flashMessenger()->setNamespace('error')->addMessage($translator->translate('The file you specified does not exist.'));
+        $this->flashMessenger()->setNamespace('error')->addMessage($this->getTranslator()->translate('The file you specified does not exist.'));
       }
     }
 
-    $this->redirectToLastPage();
+    //$this->redirectToLastPage();
   }
 
 }
